+++ src/lib/services/contractService.ts
@@ -15,6 +15,7 @@ import {
   getTxsByContractAddress,
   getRelatedProposalsCountByContractAddress,
   getRelatedProposalsByContractAddress,
+  getAdminByContractAddressesQueryDocument,
 } from "lib/data/queries";
 import type { ContractLocalInfo } from "lib/stores/contract";
 import type {
@@ -113,6 +114,33 @@ export const useInstantiateDetailByContractQuery = (
       keepPreviousData: true,
     }
   );
+};
+
+export const useAdminByContractAddresses = (
+  contractAddresses: Option<ContractAddr[]>
+): UseQueryResult<Option<Record<ContractAddr, string>>> => {
+  const queryFn = useCallback(async () => {
+    if (!contractAddresses) return undefined;
+
+    return indexerGraphClient
+      .request(getAdminByContractAddressesQueryDocument, {
+        contractAddresses,
+      })
+      .then(({ contracts }) =>
+        contracts.reduce(
+          (prev, contract) => ({
+            ...prev,
+            [contract.address as ContractAddr]: contract.account?.address,
+          }),
+          {}
+        )
+      );
+  }, [contractAddresses]);
+
+  return useQuery(["admin_by_contracts", contractAddresses], queryFn, {
+    keepPreviousData: true,
+    enabled: !!contractAddresses,
+  });
 };
 
 export const useExecuteTxsByContractAddress = (